tutorials/generate_prereq_practice/python/requirements.txt
# HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# OpenAI API for generating practice questions
openai>=1.0.0
